requirements.txt
aiohttp==3.7.4.post0
asgiref==3.5.2
astroid==2.0.4
asttokens==2.0.5
async-timeout==3.0.1
asyncpg==0.18.1
attrs==18.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2018.10.15
cffi==1.15.0
chardet==3.0.4
decorator==5.1.1
discord.py==1.7.3
Django==2.1.15
executing==0.8.3
gurobipy==9.1.0
idna==2.6
idna-ssl==1.1.0
ipython==8.4.0
isort==4.3.4
jedi==0.18.1
julia==0.1.5
lazy-object-proxy==1.3.1
markovify==0.7.1
matplotlib-inline==0.1.3
mccabe==0.6.1
multidict==6.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.19
Pygments==2.12.0
pylint==2.1.1
PyNaCl==1.2.1
python-json-logger==2.0.2
six==1.11.0
sqlparse==0.4.2
stack-data==0.2.0
traitlets==5.2.2.post1
typing_extensions==4.2.0
ujson==1.35
Unidecode==1.0.22
urllib3==1.22
wcwidth==0.2.5
websockets==6.0
wrapt==1.10.11
yarl==1.1.0
